Add a non-working need_to_draw boolean.
This boolean currently only suppresses drawing at startup.
My goal was to reduce CPU-clogging when no new images were
rendering. However, I couldn't get it to work, and I suspect
that each time love.draw() is called, it automatically
clears the drawing surface for you, so that an empty call will
actually do something, so that it's not the do-nothing operation
I was hoping for.