Fixes in zbuffer calc.
These fixes get rid of some infinite loops (oops). I have not yet
actually drawn the results to screen, so there is more testing still.

Also, I can see already that this much work is slow. Ideally, we
will have an incremental version of this code, so I'll keep that in
mind for the future. In the meantime, my medium-term goal for foocu
is a minimum engine that will get me to the point of feeling a
world, storyline, combat mechanics, and overall fun-ness. So the
rendering engine can be hacky for now.