Draw based on the zbuffer. Buggy.
I've seen several things go wrong. The first to fix is that the movement feels
jumpy when the hero goes up or down in height. I started adding some
border-drawing code, but things are _very_ slow right now. I suspect the
bottleneck may be the height computation, which is now called many times per
point with the border computation, so a return-value cache might help out.

To summarize next things:
* Actually draw the borders
* Try to speed things up by caching/memoizing map heights
* Debug the zbuffer rendering