Draw the foreground after we draw the hero.
We can now see the transparent foreground on top of the hero, but the rendering
is still far from perfect. The depth is not clear enough, and the choice of
what is transparent and how it is rendered is not great. For example, slight
overlaps to the right or left of the hero don't feel right when they're
transparent. The green borders are confusing sometimes when they are seen
through a transparent foreground sprite.

Hopefully switching to third-height slopes and back to square base tiles will
help clarify things a little. My general plan is to work through a few more
incremental rendering changes like this, then improve the map generation and
sprites. After that comes other game mechanics like conversations and combat.