Improve the falling anim; climbing crashes for now
So it might be nice to fix that crash next, and in the process implement the
climbing animation as well. I can see this will need more tweaking, as it looks
like the hero is falling through some ground. Basically, we need to do smarter
edge detection to figure out when he should climb or fall -- essentially treat
the hero as an entire sprite instead of as existing at a single point.