Fix the climb animation.
There's still a bug where if the hero falls near the bottom of the screen, or
if he climbs near the top, the map jumps around a bit. I think this is due to
some erroneous behavior now in scroll_if_needed() that maybe can be fixed by
accounting for hero_anim_offset() within that function.