Rendering no longer crashes, but is still buggy.
I'm not sure why the rendering is so far off. I plan to attack in a few steps:
1. Draw the map sprites correctly.
2. Fix the jumpy fall/climb animation.
3. Fix the way borders are drawn, including the blue/white char location ones.

For step 1, I'm thinking about checking a single column by hand. Either the
zbuffer is off, or we're drawing it wrong, or both. To check a single column by
hand, I can print out the map heights and the hero height, and print out the
zbuffer column, and check that they match.